MPI parallelization has been added: makeall-serial produces serial .exe in TDplas and WaveT (WaveT-serial.x); makeall-parallel produces all the serial .exe in TDplas and WaveT, and WaveT-parallel.x in WaveT;make-parallel allows one to compile the parallel.exe, but attention must be paid to the object files (coming from a serial or a parallel compilation). Moreover, typing "make clean_obj" delete all the object files. The codes will be preprocessed using the MPI flag. MPI parallelization is effective only when SSE is used (different seeds for each trajectory will be determined automatically).
Ionization is taken into account by the variable absorber ('y' or 'n') into the namelist general. A file ion_rate.dat is needed, of the form "index energy ionization_rate", according to the output of a modified version of Molpro, only ionization_rate will be used. Example added in tests/