Main change: matmul has been substituted with a homemade function mat_mult (and cmat_mult), which uses matmul or OMP-optimized explicit loops (LOMP) for matrix/vector multiplication. This allowed us to optimize operations which scale with the square of the dimension (number of states or number of tesserae) Modifications are in Mathtools and td_contmed (TDplas), and in propagate and dissipation (WaveT). Choice between matmul and LOMP depends on the value of Fopt (WaveT) and Fopt_chr (TDplas), which in turn depend on the value of ncit (threshold for n_ci, 150 default, in namelist general) and ntst (threshold for nts_act, 150 default, in namelist propagate):
1) use LOMP if n_ci>ncit (nts_act>ntst) AND number of threads > 1;

2) use matmul otherwise.

For a test case with 300 states and 370 states, a speedup >10 is seen (from 12h27m to 1h18m).
Threshold values are optimized to use the fastest function, according the dimension of the problem. User should not, in principle, change those values. Function mat_mult refers to matrix/vector multiplications involving the tessera space, while cmat_mult to multiplications in the state space.

By default, now, SSE calculation do not write on-the-fly information on the type of quantum jump, to switch on writing, use out_sse='y' in the namelist for SSE.

Added a directory config/, containing specific configurations for makfiles (I added those for Curie).